[TAN-8036] [TAN-8036] Consolidate Follow input-selector e2e tests into one file
