+++ front/cypress/e2e/admin/smart_group_follow_input_selectors.cy.ts
@@ -1,35 +1,48 @@
 import moment = require('moment');
 import { randomString } from '../../support/commands';
 
-// Behaviour tests for the smart-group "Follow" input value selectors, which were
-// switched from a fixed 26-random-input dropdown to a searchable, infinite-scroll
-// picker (IdeaSelect / IdeaMultiSelect). These assert what the old non-searchable
-// dropdown could not: that an arbitrary input is findable by title and selectable.
+// Behaviour + regression tests for the smart-group "Follow" input value
+// selectors. They were switched from a fixed 26-random-input dropdown to a
+// searchable, infinite-scroll picker (IdeaSingleSelect / IdeaMultiSelect), and
+// list followable inputs by fetching `/ideas`. The underlying fix removed
+// `transitive=true`, which had excluded proposals (proposals live in a proposals
+// phase, so they have a non-nil creation_phase).
 //
-// The candidate inputs are created here via the API (rather than relying on the
-// e2e template's seeded content), so the titles searched for are known and stable.
+// Coverage:
+//  - the /ideas request carries no `transitive`, so proposals are not excluded;
+//  - survey responses are never publicly visible, so are never offered as inputs;
+//  - an arbitrary input is findable by title and selectable (multi and single).
+//
+// Candidate inputs are created via the API (not the seeded e2e template) so the
+// titles searched for are known and stable.
 describe('Smart group "Follow" input value selectors', () => {
   const ideaTitleA = randomString(20);
   const ideaTitleB = randomString(20);
   const proposalTitle = randomString(20);
+  const surveyResponseTitle = randomString(20);
 
   const nineMonthsAgo = moment().subtract(9, 'month').format('DD/MM/YYYY');
   const twoMonthsAgo = moment().subtract(2, 'month').format('DD/MM/YYYY');
   const inTwoMonths = moment().add(2, 'month').format('DD/MM/YYYY');
 
-  // Two projects keeps the ideation and proposals phases on independent
-  // timelines (phases within one project cannot overlap).
+  // A separate project per phase keeps each on an independent timeline (phases
+  // within one project cannot overlap).
   let ideationProjectId: string;
   let proposalsProjectId: string;
+  let surveyProjectId: string;
+  let surveyResponseId: string;
 
-  before(() => {
-    // Ideation project + phase with two published inputs.
+  const createProject = () =>
     cy.apiCreateProject({
       title: randomString(),
       descriptionPreview: randomString(),
       description: randomString(),
       publicationStatus: 'published',
-    }).then((project) => {
+    });
+
+  before(() => {
+    // Ideation phase + two published inputs (nil creation_phase, publicly visible).
+    createProject().then((project) => {
       ideationProjectId = project.body.data.id;
       cy.apiCreatePhase({
         projectId: ideationProjectId,
@@ -52,14 +65,9 @@ describe('Smart group "Follow" input value selectors', () => {
       });
     });
 
-    // Separate project with a proposals phase + a published proposal. Proposals
+    // Proposals phase + a published proposal (non-nil creation_phase). Proposals
     // were the inputs the old `transitive=true` fetch excluded.
-    cy.apiCreateProject({
-      title: randomString(),
-      descriptionPreview: randomString(),
-      description: randomString(),
-      publicationStatus: 'published',
-    }).then((project) => {
+    createProject().then((project) => {
       proposalsProjectId = project.body.data.id;
       cy.apiCreatePhase({
         projectId: proposalsProjectId,
@@ -79,15 +87,44 @@ describe('Smart group "Follow" input value selectors', () => {
         });
       });
     });
+
+    // Native-survey phase + a survey response (creation_phase is a survey phase,
+    // so it is NOT publicly visible and must never be offered as an input).
+    createProject().then((project) => {
+      surveyProjectId = project.body.data.id;
+      cy.apiCreatePhase({
+        projectId: surveyProjectId,
+        title: randomString(),
+        startAt: twoMonthsAgo,
+        endAt: inTwoMonths,
+        participationMethod: 'native_survey',
+        nativeSurveyButtonMultiloc: { en: 'Take the survey' },
+        nativeSurveyTitleMultiloc: { en: 'Survey' },
+        canPost: true,
+        canComment: true,
+        canReact: true,
+      }).then((phase) => {
+        cy.apiCreateIdea({
+          projectId: surveyProjectId,
+          ideaTitle: surveyResponseTitle,
+          ideaContent: randomString(60),
+          phaseIds: [phase.body.data.id],
+        }).then((idea) => {
+          surveyResponseId = idea.body.data.id;
+        });
+      });
+    });
   });
 
   after(() => {
-    if (ideationProjectId) cy.apiRemoveProject(ideationProjectId);
-    if (proposalsProjectId) cy.apiRemoveProject(proposalsProjectId);
+    [ideationProjectId, proposalsProjectId, surveyProjectId].forEach((id) => {
+      if (id) cy.apiRemoveProject(id);
+    });
   });
 
   // Opens the create-smart-group modal and selects the Follow rule with the given
-  // predicate, which mounts the input value selector under test.
+  // predicate, which mounts the input value selector under test. Returns the
+  // mount-time `/ideas?…` request so callers can assert on it.
   const openFollowRule = (predicateLabel: string) => {
     // Regex (literal `?`) matches only the `/ideas?…` list requests, never the
     // `/ideas/:id` item requests the selector fires to resolve selected titles.
@@ -105,15 +142,28 @@ describe('Smart group "Follow" input value selectors', () => {
       .eq(1)
       .select(predicateLabel);
 
-    // The selector fetches its first page on mount; that request must not carry
-    // `transitive` (which would exclude proposals).
-    cy.wait('@ideaOptions')
-      .its('request.url')
-      .should('not.include', 'transitive');
+    return cy.wait('@ideaOptions');
   };
 
+  it('offers ideas and proposals as inputs, but not survey responses', () => {
+    openFollowRule('one of the inputs').then((interception) => {
+      // The selector must not restrict inputs to non-proposal ideas.
+      expect(interception.request.url).not.to.include('transitive');
+      expect(interception.response?.statusCode).to.eq(200);
+
+      // Survey responses are never publicly visible, so are never offered —
+      // robust regardless of which random page the selector received.
+      const offeredIds = interception.response?.body.data.map(
+        (idea: { id: string }) => idea.id
+      );
+      expect(offeredIds).not.to.include(surveyResponseId);
+    });
+  });
+
   it('searches and selects multiple inputs for "one of the inputs"', () => {
-    openFollowRule('one of the inputs');
+    openFollowRule('one of the inputs')
+      .its('request.url')
+      .should('not.include', 'transitive');
 
     cy.get('.e2e-rules-field-section').within(() => {
       // Find a specific input by title and select it.
@@ -137,7 +187,9 @@ describe('Smart group "Follow" input value selectors', () => {
   });
 
   it('searches and selects a single input for "excludes input"', () => {
-    openFollowRule('excludes input');
+    openFollowRule('excludes input')
+      .its('request.url')
+      .should('not.include', 'transitive');
 
     cy.get('.e2e-rules-field-section').within(() => {
       cy.get('#idea-select-input').type(ideaTitleA);
